improving preprocessing features
