+++ rinex/src/processing/mod.rs
@@ -1,0 +1,8 @@
+use qc_traits::Preprocessing;
+
+use crate::prelude::Rinex;
+
+mod decim;
+mod repair;
+
+impl Preprocessing for Rinex {}
