+++ tutorials/FILEGEN/obs-decimation.sh
@@ -2,6 +2,6 @@
 # Observation record decimation and synthesis
 
 ./target/release/rinex-cli \
-    -P decim:3h \
+    -P "decim:3 h" \
     --fp test_resources/CRNX/V3/ESBC00DNK_R_20201770000_01D_30S_MO.crx.gz \
     filegen
